Code/requirements.txt
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
Deprecated==1.2.13
dill==0.3.6
dimod==0.12.4
docplex==2.25.236
dwave-neal==0.6.0
dwave-networkx==0.8.13
dwave-samplers==1.0.0
fonttools==4.39.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==2.8.8
ntlm-auth==1.5.0
numpy==1.23.5
packaging==23.0
pandas==2.0.1
pbr==5.11.1
Pillow==9.4.0
ply==3.11
psutil==5.9.5
pycodestyle==2.10.0
pycparser==2.21
pylatexenc==2.10
pyparsing==3.0.9
pyqubo==1.4.0
python-dateutil==2.8.2
pytz==2023.3
qiskit==0.42.1
qiskit-aer==0.12.0
qiskit-ibm-provider==0.5.2
qiskit-ibmq-provider==0.20.2
qiskit-optimization==0.5.0
qiskit-terra==0.23.3
qubovert==1.2.5
requests==2.28.2
requests-ntlm==1.1.0
rustworkx==0.12.1
scipy==1.10.1
six==1.16.0
stevedore==5.0.0
sympy==1.11.1
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
websocket-client==1.5.1
websockets==11.0.2
wrapt==1.15.0
